Change playback position type from float to double
Two reasons to change this:
* At a default mix rate of 44100, the playback position in seconds can experience rounding errors with a 32-bit type if the value is high enough. 44100 requires 15 free bits in the mantissa to be within 1/2 an audio frame, so the cutoff is 512 seconds before rounding issues occur (512=2^9, and 9+15 > 23 mantissa bits in 32-bit float).
* Internally, AudioStreamPlayback and AudioStream use a 64-bit type for playback position.

See this discussion: https://github.com/godotengine/godot/pull/105510#discussion_r2049745296

+++ scene/2d/audio_stream_player_2d.cpp
@@ -253,7 +253,7 @@ void AudioStreamPlayer2D::_play_internal(double p_from_pos) {
 	}
 }
 
-void AudioStreamPlayer2D::play(float p_from_pos) {
+void AudioStreamPlayer2D::play(double p_from_pos) {
 	internal->scheduled_time = 0;
 	_play_internal(p_from_pos);
 }
@@ -263,7 +263,7 @@ void AudioStreamPlayer2D::play_scheduled(double p_abs_time, double p_from_pos) {
 	_play_internal(p_from_pos);
 }
 
-void AudioStreamPlayer2D::seek(float p_seconds) {
+void AudioStreamPlayer2D::seek(double p_seconds) {
 	internal->seek(p_seconds);
 }
 
@@ -279,7 +279,7 @@ bool AudioStreamPlayer2D::is_playing() const {
 	return internal->is_playing();
 }
 
-float AudioStreamPlayer2D::get_playback_position() {
+double AudioStreamPlayer2D::get_playback_position() {
 	if (setplay.get() >= 0) {
 		return setplay.get(); // play() has been called this frame, but no playback exists just yet.
 	}
